+++ src/bitmap.c
@@ -6,11 +6,11 @@
 Bitmap character_bitmaps[10];
 Bitmap colon;
 Bitmap blank;
-uint8_t screen_bitmap_array[SCREEN_HEIGHT][SCREEN_WIDTH]; //This should be the only bitmap we store (i think.)
+uint16_t screen_bitmap_array[SCREEN_HEIGHT][SCREEN_WIDTH]; //This should be the only bitmap we store (i think.)
 Bitmap screen_bitmap;
 
 // Bitmap definitions (fonts basically)
-const uint8_t BLANK_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t BLANK_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {0, 0, 0, 0, 0}, 
   {0, 0, 0, 0, 0},
   {0, 0, 0, 0, 0},
@@ -18,7 +18,7 @@ const uint8_t BLANK_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {0, 0, 0, 0, 0}
 };
 
-const uint8_t ZERO_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t ZERO_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255},
   {255, 0, 0, 0, 255},
   {255, 0, 0, 0, 255},
@@ -26,7 +26,7 @@ const uint8_t ZERO_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}
 };
 
-const uint8_t ONE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t ONE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {0, 0, 255, 0, 0}, 
   {0, 0, 255, 0, 0},
   {0, 0, 255, 0, 0},
@@ -34,7 +34,7 @@ const uint8_t ONE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {0, 0, 255, 0, 0}
 };
 
-const uint8_t TWO_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t TWO_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}, 
   {0, 0, 0, 0, 255},
   {255, 255, 255, 255, 255},
@@ -42,7 +42,7 @@ const uint8_t TWO_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}
 };  
 
-const uint8_t THREE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t THREE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}, 
   {0, 0, 0, 0, 255},
   {255, 255, 255, 255, 255},
@@ -50,7 +50,7 @@ const uint8_t THREE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}
 };
 
-const uint8_t FOUR_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t FOUR_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {0, 255, 255, 255, 0}, 
   {255, 0, 0, 255, 0},
   {255, 255, 255, 255, 255},
@@ -58,7 +58,7 @@ const uint8_t FOUR_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {0, 0, 0, 255, 0}
 };
 
-const uint8_t FIVE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t FIVE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}, 
   {255, 0, 0, 0, 0},
   {255, 255, 255, 255, 255},
@@ -66,7 +66,7 @@ const uint8_t FIVE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}
 };
 
-const uint8_t SIX_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t SIX_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}, 
   {255, 0, 0, 0, 0},
   {255, 255, 255, 255, 255},
@@ -74,7 +74,7 @@ const uint8_t SIX_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}
 };
 
-const uint8_t SEVEN_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t SEVEN_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}, 
   {0, 0, 0, 0, 255},
   {0, 0, 0, 255, 0},
@@ -82,8 +82,8 @@ const uint8_t SEVEN_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {0, 0, 255, 0, 0}
 };
 
-//uint8_t: create fonts from here on down.
-const uint8_t EIGHT_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+//uint16_t: create fonts from here on down.
+const uint16_t EIGHT_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   {255, 255, 255, 255, 255}, 
   {255, 0, 0, 0, 255},
   {255, 255, 255, 255, 255},
@@ -92,7 +92,7 @@ const uint8_t EIGHT_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
 };
 
 
-const uint8_t NINE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
+const uint16_t NINE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
   
   {255, 255, 255, 255, 255}, 
   {255, 0, 0, 0, 255},
@@ -102,7 +102,7 @@ const uint8_t NINE_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] = {
 };
 
 
-const uint8_t COLON_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] =  {
+const uint16_t COLON_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] =  {
   {0, 0, 0, 0, 0}, 
   {0, 0, 255, 0, 0},
   {0, 0, 0, 0, 0},
@@ -113,9 +113,9 @@ const uint8_t COLON_5X5[CHARACTER_HEIGHT][CHARACTER_WIDTH] =  {
 
 
 Bitmap create_bitmap(
-  uint8_t width, 
-  uint8_t height, 
-  const uint8_t *bitmap_loc) {
+  uint16_t width, 
+  uint16_t height, 
+  const uint16_t *bitmap_loc) {
   Bitmap character;
   // sets the character information.  
   character.width = width;
@@ -125,13 +125,13 @@ Bitmap create_bitmap(
   return character;
 }
 
-uint8_t get_character_bool_from_position(
+uint16_t get_character_bool_from_position(
   Bitmap character, 
-  uint8_t row, 
-  uint8_t col) {
+  uint16_t row, 
+  uint16_t col) {
   
-  const uint8_t offset = col + character.width * row;
-  const uint8_t *bitmap_pos = character.bitmap_loc + offset;
+  const uint16_t offset = col + character.width * row;
+  const uint16_t *bitmap_pos = character.bitmap_loc + offset;
   int bitmap_bool = *bitmap_pos;
   return bitmap_bool;
   
@@ -182,16 +182,16 @@ Bitmap get_character_bitmap(char character) {
 void overlay_bitmaps(
     Bitmap *underlay,
     Bitmap overlay,
-    uint8_t start_col, 
-    uint8_t start_row) 
+    uint16_t start_col, 
+    uint16_t start_row) 
     {
-    uint8_t pos_within_overlay = 0;
+    uint16_t pos_within_overlay = 0;
     uint16_t pos_within_underlay = start_row * SCREEN_WIDTH + start_col;
     
     while (pos_within_overlay < overlay.height * overlay.width) {
         // Copy the data BEFORE incrementing positions
-        uint8_t *underlay_pos_pointer = underlay->bitmap_loc + pos_within_underlay;
-        const uint8_t *overlay_pos_pointer = overlay.bitmap_loc + pos_within_overlay;
+        uint16_t *underlay_pos_pointer = underlay->bitmap_loc + pos_within_underlay;
+        const uint16_t *overlay_pos_pointer = overlay.bitmap_loc + pos_within_overlay;
         
         if (pos_within_underlay < SCREEN_HEIGHT * SCREEN_WIDTH) { // guard against improper assignment
             *underlay_pos_pointer = *overlay_pos_pointer;
@@ -208,8 +208,8 @@ void overlay_bitmaps(
 }
 
 
-uint8_t check_string_length(char *str_ptr) {
-  uint8_t string_length = 0;
+uint16_t check_string_length(char *str_ptr) {
+  uint16_t string_length = 0;
   char current_char = *str_ptr;
   while (current_char != '\0') {
     string_length++;
@@ -222,8 +222,8 @@ uint8_t check_string_length(char *str_ptr) {
 void update_bitmap_with_string(
   Bitmap *bitmap, 
   char *str_ptr, 
-  uint8_t start_row, 
-  uint8_t start_col) {
+  uint16_t start_row, 
+  uint16_t start_col) {
   //only works with fixed size characters for now!
   unsigned int string_length = check_string_length(str_ptr);
   char current_char = *str_ptr;
@@ -241,31 +241,31 @@ void update_bitmap_with_string(
 
 void init_fonts() {
     //defining character_bitmaps/
-    character_bitmaps[0] = create_bitmap(5, 5, (const uint8_t*)ZERO_5X5);
-    character_bitmaps[1] = create_bitmap(5, 5, (const uint8_t*)ONE_5X5);
-    character_bitmaps[2] = create_bitmap(5, 5, (const uint8_t*)TWO_5X5);
-    character_bitmaps[3] = create_bitmap(5, 5, (const uint8_t*)THREE_5X5);
-    character_bitmaps[4] = create_bitmap(5, 5, (const uint8_t*)FOUR_5X5);
-    character_bitmaps[5] = create_bitmap(5, 5, (const uint8_t*)FIVE_5X5);
-    character_bitmaps[6] = create_bitmap(5, 5, (const uint8_t*)SIX_5X5);
-    character_bitmaps[7] = create_bitmap(5, 5, (const uint8_t*)SEVEN_5X5);
-    character_bitmaps[8] = create_bitmap(5, 5, (const uint8_t*)EIGHT_5X5);
-    character_bitmaps[9] = create_bitmap(5, 5, (const uint8_t*)NINE_5X5);
+    character_bitmaps[0] = create_bitmap(5, 5, (const uint16_t*)ZERO_5X5);
+    character_bitmaps[1] = create_bitmap(5, 5, (const uint16_t*)ONE_5X5);
+    character_bitmaps[2] = create_bitmap(5, 5, (const uint16_t*)TWO_5X5);
+    character_bitmaps[3] = create_bitmap(5, 5, (const uint16_t*)THREE_5X5);
+    character_bitmaps[4] = create_bitmap(5, 5, (const uint16_t*)FOUR_5X5);
+    character_bitmaps[5] = create_bitmap(5, 5, (const uint16_t*)FIVE_5X5);
+    character_bitmaps[6] = create_bitmap(5, 5, (const uint16_t*)SIX_5X5);
+    character_bitmaps[7] = create_bitmap(5, 5, (const uint16_t*)SEVEN_5X5);
+    character_bitmaps[8] = create_bitmap(5, 5, (const uint16_t*)EIGHT_5X5);
+    character_bitmaps[9] = create_bitmap(5, 5, (const uint16_t*)NINE_5X5);
     
     //defining non-number characters(this will eventually contain whole
     //alphabet!)
-    colon = create_bitmap(5, 5, (const uint8_t*)COLON_5X5);
-    blank = create_bitmap(5, 5, (const uint8_t*)BLANK_5X5);
+    colon = create_bitmap(5, 5, (const uint16_t*)COLON_5X5);
+    blank = create_bitmap(5, 5, (const uint16_t*)BLANK_5X5);
 };
 
 void reset_screen_bitmap() {
   //set every element in screen to 0.
-  uint8_t i = 0;
-  uint8_t j = 0;
+  uint16_t i = 0;
+  uint16_t j = 0;
 
   while (i < SCREEN_HEIGHT) {
     while (j < SCREEN_WIDTH) {
-      uint8_t *screen_pixel_pointer = screen_bitmap.bitmap_loc + SCREEN_WIDTH * i + j;
+      uint16_t *screen_pixel_pointer = screen_bitmap.bitmap_loc + SCREEN_WIDTH * i + j;
       *screen_pixel_pointer = 0; 
       j++;
     }
@@ -274,6 +274,6 @@ void reset_screen_bitmap() {
 }
 
 void init_screen_bitmap() {
-  screen_bitmap = create_bitmap(SCREEN_HEIGHT, SCREEN_WIDTH, (const uint8_t*)screen_bitmap_array);
+  screen_bitmap = create_bitmap(SCREEN_HEIGHT, SCREEN_WIDTH, (const uint16_t*)screen_bitmap_array);
   reset_screen_bitmap();
 }
